media-generator-mvp/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
moviepy==1.0.3
pillow==10.1.0
numpy==1.25.2
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
TTS==0.20.6
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
diffusers==0.24.0
accelerate==0.25.0
opencv-python==4.8.1.78
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
